+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Fix
+
+* `makeCssHexColorDefinition()` no longer normalises to `##RRGGBB` format
+  * `#RGB` inputs are now preserved
+
 ## v2.0.1
 
 Released Tuesday, 25th February 2025.